pref: 构建日志模块增加服务端压缩请求 #9650
