+++ src/backend/ci/core/log/api-log/src/main/kotlin/com/tencent/devops/common/log/pojo/message/LogMessageWithLineNo.kt
@@ -56,4 +56,3 @@ data class LogMessageWithLineNo(
         executeCount = executeCount
     )
 }
-
